requirements.txt
langchain
langgraph
langchain-google-genai
yfinance
duckduckgo-search
streamlit
faiss-cpu
pandas
python-dotenv
langchain-huggingface
sentence-transformers
